bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Chroma7p <nahco3,hcl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord = "^2.3.2"
python-dotenv = "^1.0.1"
openai = "^1.17.1"
qdrant-client = "^1.8.2"
tiktoken = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
